Improve default button size on macOS
wxOSX's default was a bit too larger on at least macOS 11+.

The new value is still slightly larger than what Interface Builder
creates by default, but matches the size used in native file open
windows (on macOS 26, it is 1pt larger on macOS 15, but that's not
noticeable and not worth complicating the code for).

Note that native AppKit apps often use even smaller button sizes.