Guard macOS 26 checks with __MAC_OS_X_VERSION_MAX_ALLOWED
Wrap uses of WX_IS_MACOS_AVAILABLE(26, 0) with compile-time
__MAC_OS_X_VERSION_MAX_ALLOWED >= MAC_OS_VERSION_26_0 checks so that
code compiled with older versions of Xcode 26 is not affects.

macOS 26 doesn't apply its new appearance to applications that weren't
compiled with Xcode 26+, so this is the correct thing to do: we do not
want any adjustments to apply to 26-unaware code.